requirements.txt
asgiref==3.5.2
Django==4.1.2
django-autoslug==1.9.8
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-writable-nested==0.7.0
drf-yasg==1.21.4
gunicorn==20.1.0
python-dotenv==0.21.0
PyYAML==6.0
uritemplate==4.1.1
python-decouple==3.7